% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_coagulation.R
\name{score_coagulation}
\alias{score_coagulation}
\title{Coagulation component of the SOFA score}
\usage{
score_coagulation(platelets)
}
\arguments{
\item{platelets}{Numeric vector of platelet counts in \code{10^9 / L}
(equivalent to \code{1000 / mm^3}).}
}
\value{
Integer vector in \verb{\{0, 1, 2, 3, 4\}} (or \code{NA_integer_}).
}
\description{
Scores the coagulation subscale of the SOFA from the platelet count.
}
\section{Scoring rules}{

\itemize{
\item \verb{>= 150}: 0
\item \verb{< 150}:  1
\item \verb{< 100}:  2
\item \verb{<  50}:  3
\item \verb{<  20}:  4
}
}

\examples{
score_coagulation(c(180, 130, 75, 35, 12, NA))

}
